+++ source/common/protobuf/message_validator_impl.h
@@ -21,5 +21,21 @@ public:
 
 ValidationVisitor& getStrictValidationVisitor();
 
+class ValidationContextImpl : public ValidationContext {
+public:
+  ValidationContextImpl(ValidationVisitor& static_validation_visitor,
+                        ValidationVisitor& dynamic_validation_visitor)
+      : static_validation_visitor_(static_validation_visitor),
+        dynamic_validation_visitor_(dynamic_validation_visitor) {}
+
+  // Envoy::ProtobufMessage::ValidationContext
+  ValidationVisitor& staticValidationVisitor() override { return static_validation_visitor_; }
+  ValidationVisitor& dynamicValidationVisitor() override { return dynamic_validation_visitor_; }
+
+private:
+  ValidationVisitor& static_validation_visitor_;
+  ValidationVisitor& dynamic_validation_visitor_;
+};
+
 } // namespace ProtobufMessage
 } // namespace Envoy