+++ source/server/options_impl.cc
@@ -51,6 +51,9 @@ OptionsImpl::OptionsImpl(int argc, const char* const* argv,
 
   TCLAP::SwitchArg allow_unknown_fields("", "allow-unknown-fields",
                                         "allow unknown fields in the configuration", cmd, false);
+  TCLAP::SwitchArg reject_unknown_fields_dynamic("", "reject-unknown-fields-dynamic",
+                                                 "reject unknown fields in dynamic configuration",
+                                                 cmd, false);
 
   TCLAP::ValueArg<std::string> admin_address_path("", "admin-address-path", "Admin address path",
                                                   false, "", "string", cmd);
@@ -182,6 +185,7 @@ OptionsImpl::OptionsImpl(int argc, const char* const* argv,
   config_path_ = config_path.getValue();
   config_yaml_ = config_yaml.getValue();
   allow_unknown_fields_ = allow_unknown_fields.getValue();
+  reject_unknown_fields_dynamic_ = reject_unknown_fields_dynamic.getValue();
   admin_address_path_ = admin_address_path.getValue();
   log_path_ = log_path.getValue();
   restart_epoch_ = restart_epoch.getValue();
@@ -242,6 +246,7 @@ Server::CommandLineOptionsPtr OptionsImpl::toCommandLineOptions() const {
   command_line_options->set_config_path(configPath());
   command_line_options->set_config_yaml(configYaml());
   command_line_options->set_allow_unknown_fields(allow_unknown_fields_);
+  command_line_options->set_reject_unknown_fields_dynamic(reject_unknown_fields_dynamic_);
   command_line_options->set_admin_address_path(adminAddressPath());
   command_line_options->set_component_log_level(component_log_level_str_);
   command_line_options->set_log_level(spdlog::level::to_string_view(logLevel()).data(),