+++ test/integration/integration.h
@@ -190,9 +190,12 @@ public:
   void registerTestServerPorts(const std::vector<std::string>& port_names);
   void createTestServer(const std::string& json_path, const std::vector<std::string>& port_names);
   void createGeneratedApiTestServer(const std::string& bootstrap_path,
-                                    const std::vector<std::string>& port_names);
+                                    const std::vector<std::string>& port_names,
+                                    bool allow_unknown_fields, bool reject_unknown_fields_dynamic,
+                                    bool allow_lds_rejection);
   void createApiTestServer(const ApiFilesystemConfig& api_filesystem_config,
-                           const std::vector<std::string>& port_names);
+                           const std::vector<std::string>& port_names, bool allow_unknown_fields,
+                           bool reject_unknown_fields_dynamic, bool allow_lds_rejection);
 
   Event::TestTimeSystem& timeSystem() { return time_system_; }